% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{apatite}
\alias{apatite}
\title{apatite compositions}
\description{
Varietal dataset of detrital apatite compositions, provided by
 L. Caracciolo (FAU Erlangen).
}
\examples{
plot(MDS(apatite))
}
