% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{as.acomp}
\alias{as.acomp}
\title{create an \code{acomp} object}
\usage{
as.acomp(x)
}
\arguments{
\item{x}{an object of class \code{compositional}}
}
\value{
a \code{data.frame}
}
\description{
Convert an object of class \code{compositional} to an object of
class \code{acomp} for use in the \code{compositions} package
}
\examples{
data(Namib)
qfl <- ternary(Namib$PT,c('Q'),c('KF','P'),c('Lm','Lv','Ls'))
plot(qfl,type="QFL.dickinson")
qfl.acomp <- as.acomp(qfl)
## uncomment the next two lines to plot an error
## ellipse using the compositions package: 
# library(compositions)
# ellipses(mean(qfl.acomp),var(qfl.acomp),r=2)
}
