% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{as.data.frame.compositional}
\alias{as.data.frame.compositional}
\title{create a \code{data.frame} object}
\usage{
\method{as.data.frame}{compositional}(x, ...)
}
\arguments{
\item{x}{an object of class \code{compositional}}

\item{...}{optional arguments to be passed on to the generic function}
}
\value{
a \code{data.frame}
}
\description{
Convert an object of class \code{compositional} to a
\code{data.frame} for use in the \code{robCompositions} package
}
\examples{
data(Namib)
Major.frame <- as.data.frame(Namib$Major)
## uncomment the next two lines to plot an error
## ellipse using the robCompositions package:
# library(robCompositions)
# plot(pcaCoDa(Major.frame))
}
