% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{as.data.frame.counts}
\alias{as.data.frame.counts}
\title{create a \code{data.frame} object}
\usage{
\method{as.data.frame}{counts}(x, ...)
}
\arguments{
\item{x}{an object of class \code{counts}}

\item{...}{optional arguments to be passed on to the generic function}
}
\value{
a \code{data.frame}
}
\description{
Convert an object of class \code{counts} to a \code{data.frame} for
use in the \code{robCompositions} package
}
\examples{
data(Namib)
qfl <- ternary(Namib$PTHM,c('Q'),c('KF','P'),c('Lm','Lv','Ls'))
plot(qfl,type="QFL.dickinson")
qfl.frame <- as.data.frame(qfl)
## uncomment the next two lines to plot an error
## ellipse using the robCompositions package:
# library(robCompositions)
# pca <- pcaCoDa(qfl.frame)
# plot(pca,xlabs=rownames(qfl.frame))
}
