% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/provenance.R
\name{bray.diss}
\alias{bray.diss}
\title{Bray-Curtis dissimilarity}
\usage{
bray.diss(x, y)
}
\arguments{
\item{x}{a vector containing the first compositional sample}

\item{y}{a vector of length(x) containing the second compositional sample}
}
\value{
a scalar value
}
\description{
Calculates the Bray-Curtis dissimilarity between two samples
}
\examples{
data(Namib)
print(bray.diss(Namib$HM$x["N1",],Namib$HM$x["N2",]))
}

