% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{CA}
\alias{CA}
\title{Correspondence Analysis}
\usage{
CA(x, nf = 2, ...)
}
\arguments{
\item{x}{an object of class \code{counts}}

\item{nf}{number of correspondence factors (dimensions)}

\item{...}{optional arguments to the \code{corresp} function of the
\code{MASS} package}
}
\value{
an object of classes \code{CA}, which is synonymous to the
    MASS package's \code{correspondence} class.
}
\description{
Performs Correspondence Analysis of point-counting data
}
\examples{
data(Namib)
plot(CA(Namib$PT))
}
