% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radialplot.R
\name{central}
\alias{central}
\title{Calculate central compositions}
\usage{
central(x, components = NA, alpha = 0.05, ...)
}
\arguments{
\item{x}{an object of class \code{counts}}

\item{components}{a vector of column names defining a
subcomposition}

\item{alpha}{cutoff value for confidence intervals}

\item{...}{optional arguments}
}
\value{
a list containing:

\describe{
\item{mswd}{ the reduced chi-square statistic for a simple fit. }
\item{p.value}{ the p-value for age homogeneity. }
\item{ratio}{ the central ratio. }
\item{err}{ the standard error for the central ratio. }
\item{sigma}{ the overdispersion parameter, i.e. the coefficient of variation of the
underlying logistic normal distribution. }
}
}
\description{
Computes the geometric mean composition of a continuous mixture of
point-counting data.
}
\details{
The central composition assumes that the observed point-counting
distribution is the combination of two sources of scatter:
counting uncertainty and true geological dispersion.
}
