% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarities.R
\name{diss}
\alias{diss}
\alias{diss.distributional}
\alias{diss.compositional}
\alias{diss.counts}
\alias{diss.varietal}
\title{Calculate the dissimilarity matrix between two datasets of class
\code{distributional}, \code{compositional}, \code{counts} or
\code{varietal}}
\usage{
diss(x, method, ...)

\method{diss}{distributional}(x, method = NULL, log = FALSE, ...)

\method{diss}{compositional}(x, method = NULL, ...)

\method{diss}{counts}(x, method = NULL, ...)

\method{diss}{varietal}(x, method = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{distributional},
\code{compositional} or \code{counts}}

\item{method}{(optional) either "KS", "Wasserstein", "Kuiper",
"SH", "aitchison", "bray" or "chisq"}

\item{...}{optional arguments}

\item{log}{logical. If \code{TRUE}, subjects the distributional
data to a logarithmic transformation before calculating the
Wasserstein distance.}
}
\value{
an object of class \code{diss}
}
\description{
Calculate the dissimilarity matrix between two datasets of class
\code{distributional} or \code{compositional} using the Kolmogorov-Smirnov,
Sircombe-Hazelton, Aitchison or Bray-Curtis distance
}
\examples{
data(Namib)
print(round(100*diss(Namib$DZ)))
}
