% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provenance.R
\name{get.f}
\alias{get.f}
\title{Calculate the largest fraction that is likely to be missed}
\usage{
get.f(n, p = 0.05)
}
\arguments{
\item{n}{the number of grains in the detrital sample}

\item{p}{the required level of confidence}
}
\value{
the largest fraction that is sampled with at least 100 x p%
certainty
}
\description{
For a given sample size, returns the largest fraction which has
been sampled with p x 100 % likelihood.
}
\examples{
print(get.f(60))
print(get.f(117))
}
\references{
Vermeesch, Pieter. "How many grains are needed for a
provenance study?." Earth and Planetary Science Letters 224.3
(2004): 441-451.
}

