% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/provenance.R
\name{GPA}
\alias{GPA}
\title{Generalised Procrustes Analysis of configurations}
\usage{
GPA(X, scale = TRUE)
}
\arguments{
\item{X}{a list of dissimilarity matrices}

\item{scale}{boolean flag indicating if the transformation should include the scaling operation}
}
\value{
a two column vector with the coordinates of the
group configuration
}
\description{
Given a number of (2D) configurations, this function uses a
combination of transformations (reflections, rotations,
translations and scaling) to find a 'consensus' configuration which
best matches all the component configurations in a least-squares
sense.
}
\seealso{
procrustes
}

