\name{kdemds-package}
\alias{kdemds-package}
\alias{kdemds}
\docType{package}
\title{
Plotting KDEs and MDS maps of geochronology and geochemistry data
}
\description{
This package provides helper functions to facilitate plotting kernel density estimates (KDEs) of geochronological data, and multi-dimensional scaling (MDS) maps of these and other data. These types of plots serve as a powerful tool for comparison of large datasets of e.g. zircon U-Pb age spectra (see examples).
Plotting makes heavy use of the package ggplot2, the produced plot objects can be used and modified with ggplot's functions.
}
\details{
\tabular{ll}{
Package: \tab kdemds\cr
Type: \tab Package\cr
Version: \tab 0.2-4\cr
Date: \tab 2014-03-12\cr
License: \tab GPL (>=2)\cr
}
%% ~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Martin Rittner (2013-2014)

Maintainer: Martin Rittner <sw@thegeologician.net>
}
\references{
%% ~~ Literature or other references for background information ~~
}
%% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
