% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{KS.diss}
\alias{KS.diss}
\title{Kolmogorov-Smirnov dissimilarity}
\usage{
KS.diss(x, y)
}
\arguments{
\item{x}{the first sample as a vector}

\item{y}{the second sample as a vector}
}
\value{
a scalar value representing the maximum vertical distance
between the two cumulative distributions
}
\description{
Returns the Kolmogorov-Smirnov dissimilarity between two samples
}
\examples{
data(Namib)
print(KS.diss(Namib$DZ$x[['N1']],Namib$DZ$x[['T8']]))
}
