% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/provenance.R
\name{KS.dist}
\alias{KS.dist}
\title{Kolmogorov-Smirnov distance}
\usage{
KS.dist(x, y)
}
\arguments{
\item{x}{the first sample as a vector}

\item{y}{the second sample as a vector}
}
\value{
a scalar value representing the maximum vertical distance
between the two cumulative distributions
}
\description{
Returns the Kolmogorov-Smirnov distance between two samples
}
\examples{
fname <- system.file("DZ.csv",package="provenance")
DZ <- read.DZdata(fname)
print(KS.dist(DZ$x[['N1']],DZ$x[['T8']]))
}

