% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarities.R
\name{Kuiper.diss}
\alias{Kuiper.diss}
\alias{Kuiper.diss.default}
\alias{Kuiper.diss.distributional}
\title{Kuiper dissimilarity}
\usage{
Kuiper.diss(x, ...)

\method{Kuiper.diss}{default}(x, y, ...)

\method{Kuiper.diss}{distributional}(x, ...)
}
\arguments{
\item{x}{the first sample as a vector}

\item{...}{optional arguments}

\item{y}{the second sample as a vector}
}
\value{
a scalar value representing the sum of the maximum vertical
distances above and below the cumulative distributions of x and y
}
\description{
Returns the Kuiper dissimilarity between two samples
}
\examples{
data(Namib)
print(Kuiper.diss(Namib$DZ$x[['N1']],Namib$DZ$x[['T8']]))
}
