% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{PCA}
\alias{PCA}
\title{Principal Component Analysis}
\usage{
PCA(x, ...)
}
\arguments{
\item{x}{an object of class \code{compositional}}

\item{...}{optional arguments to R's \code{princomp} function}
}
\value{
an object of classes \code{PCA}, which is synonymous to the
    stats packages' \code{princomp} class.
}
\description{
Performs PCA of compositional data using a centred logratio
distance
}
\examples{
data(Namib)
plot(MDS(Namib$Major,classical=TRUE))
dev.new()
plot(PCA(Namib$Major),asp=1)
print("This example demonstrates the equivalence of classical MDS and PCA")
}
