% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.CA}
\alias{plot.CA}
\title{Point-counting biplot}
\usage{
\method{plot}{CA}(x, ...)
}
\arguments{
\item{x}{an object of class \code{CA}}

\item{...}{optional arguments of the \code{biplot} function}
}
\description{
Plot the results of a correspondence analysis as a biplot
}
\examples{
data(Namib)
plot(CA(Namib$PT))
}
\seealso{
CA
}
