% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.compositional}
\alias{plot.compositional}
\title{Plot a pie chart}
\usage{
\method{plot}{compositional}(x, sname, annotate = TRUE, colmap = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{compositional}}

\item{sname}{the sample name}

\item{annotate}{a boolean flag controlling if the pies of the
pie-chart should be labelled}

\item{colmap}{an optional string with the name of one of R's
built-in colour palettes (e.g., heat.colors, terrain.colors,
topo.colors, cm.colors), which are to be used for plotting the
data.}

\item{...}{optional parameters to be passed on to the graphics
object}
}
\description{
Plots an object of class \code{compositional} as a pie chart
}
\examples{
data(Namib)
plot(Namib$Major,'N1',colmap='heat.colors')
}
