% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/provenance.R
\name{plot.GPA}
\alias{plot.GPA}
\title{Plot a Procrustes configuration}
\usage{
\method{plot}{GPA}(x, ...)
}
\arguments{
\item{x}{an object of class \code{GPA}}

\item{...}{optional arguments to the generic \code{plot} function}
}
\description{
Plots the group configuration of a Generalised Procrustes Analysis
}
\examples{
DZ <- read.DZdata(system.file("DZ.csv",package="provenance"))
HM <- read.HMdata(system.file("HM.csv",package="provenance"))
GPA <- procrustes(DZ,HM)
plot(GPA)
}

