% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/provenance.R
\name{plot.HMdata}
\alias{plot.HMdata}
\title{Plot a pie chart}
\usage{
\method{plot}{HMdata}(x, sname, annotate = TRUE, col = "default", ...)
}
\arguments{
\item{x}{an object of class \code{HMdata}}

\item{sname}{the sample name}

\item{annotate}{a boolean flag controlling if the pies of the
pie-chart should be labeled}

\item{col}{colour scheme to fill the pie charts}

\item{...}{optional parameters to be passed on to the graphics object}
}
\description{
Plots an object of class \code{HMdata} as a pie chart
}
\examples{
fname <- system.file("HM.csv",package="provenance")
HM <- read.HMdata(fname)
plot(HM,'N1')
}

