% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.KDEs}
\alias{plot.KDEs}
\title{Plot one or more kernel density estimates}
\usage{
\method{plot}{KDEs}(x, sname = NA, annotate = TRUE, pch = "|", ...)
}
\arguments{
\item{x}{an object of class \code{KDEs}}

\item{sname}{optional sample name. If \code{sname=NA}, all samples
are shown on a summary plot}

\item{annotate}{add a time axis?}

\item{pch}{symbol to be used to mark the sample points along the
x-axis. Change to \code{NA} to omit.}

\item{...}{optional parameters to be passed on to the
\code{summaryplot} function}
}
\description{
Plots an object of class \code{KDEs}
}
\examples{
data(Namib)
kdes <- KDEs(Namib$DZ)
plot(kdes,ncol=2)
}
\seealso{
KDEs summaryplot
}
