% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/provenance.R
\name{plot.PCA}
\alias{plot.PCA}
\title{Compositional biplot}
\usage{
\method{plot}{PCA}(x, ...)
}
\arguments{
\item{x}{an object of class \code{PCA}}

\item{...}{optional arguments of the \code{biplot} function}
}
\description{
Plot the results of a principal components analysis as a biplot
}
\examples{
fname <- system.file("Major.csv",package="provenance")
Major <- read.HMdata(fname)
plot(getPCA(Major))
}

