% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/provenance.R
\name{read.HMdata}
\alias{read.HMdata}
\title{Read a .csv file with categorical data}
\usage{
read.HMdata(fname, metric = NULL)
}
\arguments{
\item{fname}{a string with the path to the .csv file}

\item{metric}{either "bray" (for the Bray-Curtis distance) or
"aitchison" (for Aitchison's central logratio distance). If
omitted, the function defaults to 'aitchison', unless there are
zeros present in the data.}
}
\value{
an object of class \code{HMdata}, i.e. a list with the
following items:

\code{x}: a data frame with the samples as rows and the categories as columns

\code{metric}: either "aitchison" (for Aitchison's centred logratio
distance) or "bray" (for the Bray-Curtis distance)
}
\description{
Reads a data table containing categorical data (e.g. petrographic,
heavy mineral or geochemical data)
}
\examples{
fname <- system.file("Major.csv",package="provenance")
Major <- read.HMdata(fname)
plot(getPCA(Major))
}

