% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{SNSM}
\alias{SNSM}
\title{apatite compositions}
\description{
A list of varietal dataset of detrital zircon (\code{zr}), apatite
 (\code{ap}) and titanite (\code{tit}) compositions from the Sierra
 Nevada de Santa Marta, provided by L. Caracciolo (FAU Erlangen).
}
\examples{
plot(indscal(zircon=SNSM$zr,apatite=SNSM$ap,titanite=SNSM$tit))
}
