% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{subset.compositional}
\alias{subset.compositional}
\title{Get a subset of compositional data}
\usage{
\method{subset}{compositional}(x, subset = NULL, select = NULL,
  components = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{compositional}}

\item{subset}{logical expression indicating elements or rows to keep:
missing values are taken as false.}

\item{select}{a vector of sample names.}

\item{components}{a vector specifying a subcomposition}

\item{...}{optional arguments for the generic subset function}
}
\value{
an object of class \code{compositional}
}
\description{
Return a subset of provenance data according to some specified indices
}
\seealso{
read.compositional
}
