% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{subset.distributional}
\alias{subset.distributional}
\title{Get a subset of distributional data}
\usage{
\method{subset}{distributional}(x, subset = NULL, select = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{distributional}}

\item{subset}{logical expression indicating elements or rows to keep:
missing values are taken as false.}

\item{select}{a vector of sample names}

\item{...}{optional arguments for the generic subset function}
}
\value{
an object of class \code{distributional}
}
\description{
Return a subset of provenance data according to some specified indices
}
\examples{
data(Namib)
coast <- subset(Namib$HM,select=c("N1","N2","T8","T13","N12","N13"))
summaryplot(coast,ncol=2)
}
\seealso{
read.distributional
}
