% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ternary.ellipse}
\alias{ternary.ellipse}
\title{Ternary confidence ellipse}
\usage{
ternary.ellipse(x, alpha = 0.05, population = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{ternary}}

\item{alpha}{cutoff level for the confidence ellipse}

\item{population}{show the standard deviation of the entire
population or the standard error of the mean?}

\item{...}{optional formatting arguments}
}
\description{
plot a logistic \eqn{100(1-alpha)\%} confidence region around the
    data or around its mean.
}
