% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{titanite}
\alias{titanite}
\title{titanite compositions}
\description{
Varietal dataset of detrital titanite compositions, provided by
 L. Caracciolo (FAU Erlangen).
}
\examples{
plot(MDS(titanite))
}
