% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{data}
\name{zircon}
\alias{zircon}
\title{zircon compositions}
\description{
Varietal dataset of detrital zircon compositions, provided by
 L. Caracciolo (FAU Erlangen).
}
\examples{
plot(MDS(zircon))
}
