% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{as.varietal}
\alias{as.varietal}
\title{create a \code{varietal} object}
\usage{
as.varietal(x, snames = NULL)
}
\arguments{
\item{x}{an object of class \code{matrix} or \code{data.frame}}

\item{snames}{either a vector of sample names, an integer marking
the length of the sample name prefix, or
\code{NULL}. \code{read.varietal} assumes that the row names of
the \code{.csv} file consist of character strings marking the
sample names, followed by a number.}
}
\value{
an object of class \code{varietal}
}
\description{
Convert an object of class \code{matrix} or \code{data.frame} to an
object of class \code{varietal}
}
\examples{
data(Namib)
PT.acomp <- as.acomp(Namib$PT)
PT.compositional <- as.compositional(PT.acomp)
print(Namib$PT$x - PT.compositional$x)
## uncomment the following lines for an illustration of using this 
## function to integrate 'provenance' with 'compositions'
# library(compositions)
# data(Glacial)
# a.glac <- acomp(Glacial)
# c.glac <- as.compositional(a.glac)
# summaryplot(c.glac,ncol=8)
}
