% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{MDS}
\alias{MDS}
\alias{MDS.default}
\alias{MDS.compositional}
\alias{MDS.counts}
\alias{MDS.distributional}
\alias{MDS.diss}
\title{Multidimensional Scaling}
\usage{
MDS(x, ...)

\method{MDS}{default}(x, fn, classical = FALSE, k = 2, ...)

\method{MDS}{compositional}(x, classical = FALSE, k = 2, ...)

\method{MDS}{counts}(x, classical = FALSE, k = 2, ...)

\method{MDS}{distributional}(x, classical = FALSE, k = 2,
  bootstrap = FALSE, nb = 10, ...)

\method{MDS}{diss}(x, classical = FALSE, k = 2, ...)
}
\arguments{
\item{x}{an object of class \code{distributional},
\code{compositional}, \code{counts} or \code{diss}}

\item{...}{optional arguments to be passed onto \code{cmdscale} (if
\code{classical=TRUE}) or \code{isoMDS} (if
\code{classical=FALSE}).}

\item{fn}{a dissimilarity function. One of either
\code{diss.compositional}, \code{diss.counts}, or
\code{diss.distributional}}

\item{classical}{boolean flag indicating whether classical
(\code{TRUE}) or nonmetric (\code{FALSE}) MDS should be used}

\item{k}{the desired dimensionality of the solution}

\item{bootstrap}{resample the data to calculate confidence polygons
for the MDS configuration?}

\item{nb}{number of bootstrap resamples}
}
\value{
an object of class \code{MDS}, i.e. a list containing the
    following items:

\code{points}: a two column vector of the fitted configuration

\code{classical}: a boolean flag indicating whether the MDS
configuration was obtained by classical (\code{TRUE}) or nonmetric
(\code{FALSE}) MDS.

\code{diss}: the dissimilarity matrix used for the MDS analysis

\code{stress}: (only if \code{classical=TRUE}) the final stress
achieved (in percent)
}
\description{
Performs classical or nonmetric Multidimensional Scaling analysis
of provenance data
}
\examples{
data(Namib)
plot(MDS(Namib$Major,classical=TRUE))
}
