% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.CA}
\alias{plot.CA}
\title{Point-counting biplot}
\usage{
\method{plot}{CA}(x, labelcol = "black", vectorcol = "red", components = c(1, 2), ...)
}
\arguments{
\item{x}{an object of class \code{CA}}

\item{labelcol}{colour of the sample labels (may be a vector).}

\item{vectorcol}{colour of the vector loadings for the variables}

\item{components}{two-element vector of components to be plotted}

\item{...}{optional arguments of the generic \code{biplot} function}
}
\description{
Plot the results of a correspondence analysis as a biplot
}
\examples{
data(Namib)
plot(CA(Namib$PT))
}
\seealso{
CA
}
