% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.CA}
\alias{plot.CA}
\title{Point-counting biplot}
\usage{
\method{plot}{CA}(x, levels = NULL, labelcol = c("black", "blue"),
  vectorcol = "red", ...)
}
\arguments{
\item{x}{an object of class \code{CA}}

\item{levels}{a vector of length \code{nrow(dat$x)} with values
that are to be used for the colour scale}

\item{labelcol}{two element vector with the colours that are to be
assigned to the labels of the samples corresponding to the
minimum and maximum value of \code{levels}. If
\code{levels==NULL}, then the first item of the vector is used
for all sample labels.}

\item{vectorcol}{colour of the vector loadings for the variables}

\item{...}{optional arguments of the generic
\code{\link[stats]{biplot}} function}
}
\description{
Plot the results of a correspondence analysis as a biplot
}
\examples{
data(Namib)
plot(CA(Namib$PT))
}
\seealso{
CA
}
