% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.GPA}
\alias{plot.GPA}
\title{Plot a Procrustes configuration}
\usage{
\method{plot}{GPA}(x, ...)
}
\arguments{
\item{x}{an object of class \code{GPA}}

\item{...}{optional arguments to the generic \code{plot} function}
}
\description{
Plots the group configuration of a Generalised Procrustes Analysis
}
\examples{
data(Namib)
GPA <- procrustes(Namib$DZ,Namib$HM)
plot(GPA)
}
\seealso{
procrustes
}

