% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/provenance.R
\name{plot.HMdata}
\alias{plot.HMdata}
\title{Plot a pie chart}
\usage{
\method{plot}{HMdata}(x, sname, annotate = TRUE, colmap = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{HMdata}}

\item{sname}{the sample name}

\item{annotate}{a boolean flag controlling if the pies of the
pie-chart should be labeled}

\item{colmap}{an optional string with the name of one of R's
built-in colour palettes (e.g., heat.colors, terrain.colors,
topo.colors, cm.colors), which are to be used for plotting the data.}

\item{...}{optional parameters to be passed on to the graphics object}
}
\description{
Plots an object of class \code{HMdata} as a pie chart
}
\examples{
fname <- system.file("HM.csv",package="provenance")
HM <- read.HMdata(fname)
plot(HM,'N1',colmap='heat.colors')
}

