% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.MDS}
\alias{plot.MDS}
\title{Plot an MDS configuration}
\usage{
\method{plot}{MDS}(x, nnlines = FALSE, xlab = "", ylab = "", xaxt = "n",
  yaxt = "n", ...)
}
\arguments{
\item{x}{an object of class \code{MDS}}

\item{nnlines}{if TRUE, draws nearest neighbour lines}

\item{xlab}{a string with the label of the x axis}

\item{ylab}{a string with the label of the y axis}

\item{xaxt}{if = 's', adds ticks to the x axis}

\item{yaxt}{if = 's', adds ticks to the y axis}

\item{...}{optional arguments to the generic \code{plot} function}
}
\description{
Plots the coordinates of a multidimensional scaling analysis as an
X-Y scatter plot or 'map' and, if x$classical = FALSE, a Shepard
plot.
}
\seealso{
MDS
}

