% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.PCA}
\alias{plot.PCA}
\title{Compositional biplot}
\usage{
\method{plot}{PCA}(x, levels = NULL, labelcol = c("black", "blue"),
  vectorcol = "red", choices = 1L:2L, scale = 1, pc.biplot = FALSE,
  ...)
}
\arguments{
\item{x}{an object of class \code{PCA}}

\item{levels}{a vector of length \code{nrow(dat$x)} with values
that are to be used for the colour scale}

\item{labelcol}{two element vector with the colours that are to be
assigned to the labels of the samples corresponding to the
minimum and maximum value of \code{levels}. If
\code{levels==NULL}, then the first item of the vector is used
for all sample labels.}

\item{vectorcol}{colour of the vector loadings for the variables}

\item{choices}{see the help pages of the
\code{\link[stats]{biplot}} function.}

\item{scale}{see the help pages of the \code{\link[stats]{biplot}}
function.}

\item{pc.biplot}{see the help pages of the
\code{\link[stats]{biplot}} function.}

\item{...}{optional arguments of the generic
\code{\link[stats]{biplot}} function}
}
\description{
Plot the results of a principal components analysis as a biplot
}
\examples{
data(Namib)
plot(PCA(Namib$Major))
}
\seealso{
PCA
}
