% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.PCA}
\alias{plot.PCA}
\title{Compositional biplot}
\usage{
\method{plot}{PCA}(x, ...)
}
\arguments{
\item{x}{an object of class \code{PCA}}

\item{...}{optional arguments of the \code{biplot} function}
}
\description{
Plot the results of a principal components analysis as a biplot
}
\examples{
data(Namib)
plot(PCA(Namib$Major))
}
\seealso{
PCA
}

