% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{points.ternary}
\alias{points.ternary}
\title{Add points on a ternary diagram}
\usage{
\method{points}{ternary}(x, ...)
}
\arguments{
\item{x}{an object of class \code{ternary}, or a three-column data
frame or matrix}

\item{...}{optional arguments to the generic \code{points} function}
}
\description{
Add points on a ternary diagram
}
