% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\docType{package}
\name{provenance}
\alias{provenance}
\alias{provenance-package}
\title{library(provenance)}
\description{
'provenance' is a package for the statistical treatment of large
datasets in sedimentary provenance analysis.
}
\details{
To get detailed 'getting started' introduction use help.start() or
help.start(browser='myfavouritebrowser') go to 'Packages',
'provenance', 'User guides, package vignettes and other
documentation' and then launch the file 'provenance.pdf' from
there. Please also check the website:
http://provenance.london-geochron.com for further details and
updates.
}
\examples{
library('provenance')
data(Namib)
attach(Namib)
summaryplot(KDEs(DZ),HM,PT,Major,Trace)
}
\author{
Pieter Vermeesch
}
\references{
Vermeesch, P., Resentini, A. and Garzanti, E., an R
package for statistical provenance analysis, Sedimentary Geology
(in review).
}

