% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{provenance}
\alias{provenance}
\title{Menu-based interface for \code{provenance}}
\usage{
provenance()
}
\description{
For those less familiar with the syntax of the \code{R}
programming language, the \code{provenance()} function provides
a user-friendly way to access the most important functionality
in the form of a menu-based query interface. Further details
and examples are provided on \url{http://provenance.london-geochron.com}
}
\author{
Pieter Vermeesch
}
\references{
Vermeesch, P., Resentini, A. and Garzanti, E., an R
package for statistical provenance analysis, Sedimentary Geology,
doi:10.1016/j.sedgeo.2016.01.009.
}
\seealso{
\url{http://provenance.london-geochron.com}
}

