% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{summaryplot}
\alias{summaryplot}
\title{Joint plot of several provenance datasets}
\usage{
summaryplot(..., ncol = 1, pch = NA)
}
\arguments{
\item{...}{a sequence of datasets of class \code{compositional},
\code{distributional}, \code{counts} or \code{KDEs}.}

\item{ncol}{the number of columns}

\item{pch}{(optional) symbol to be used to mark the sample points
along the x-axis of the KDEs (if appropriate).}
}
\value{
a summary plot of all the data comprised of KDEs for the
    datasets of class \code{KDEs}, pie charts for those of class
    \code{compositional} or \code{counts} and histograms for those
    of class \code{distributional}.
}
\description{
Arranges kernel density estimates and pie charts in a grid format
}
\examples{
data(Namib)
KDEs <- KDEs(Namib$DZ,0,3000)
summaryplot(KDEs,Namib$HM,Namib$PT,ncol=2)
}
\seealso{
\link{KDEs}
}
