% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarities.R
\name{Wasserstein.diss}
\alias{Wasserstein.diss}
\alias{Wasserstein.diss.default}
\alias{Wasserstein.diss.distributional}
\title{Wasserstein distance}
\usage{
Wasserstein.diss(x, ...)

\method{Wasserstein.diss}{default}(x, y, ...)

\method{Wasserstein.diss}{distributional}(x, log = FALSE, ...)
}
\arguments{
\item{x}{the first sample as a vector}

\item{...}{optional arguments}

\item{y}{the second sample as a vector}

\item{log}{logical. Take the lograthm of the data before
calculating the distances?}
}
\value{
a scalar value
}
\description{
Returns the Wasserstein distance between two samples
}
\examples{
data(Namib)
print(Wasserstein.diss(Namib$DZ$x[['N1']],Namib$DZ$x[['T8']]))

\dontrun{
# the following code requires the 'approxOT' package:
fn <- system.file('varietalWasserstein.R',package='provenance')
source(fn)
d <- Wasserstein.diss(SNSM$tit,package='approxOT')
conf <- MDS(d)
plot(conf)

# alternatively, using the 'transport' package:
d <- Wasserstein.diss(SNSM$tit,package='transport')
plot(MDS(d))
}
}
