% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transport.R
\name{Wasserstein.diss}
\alias{Wasserstein.diss}
\alias{Wasserstein.diss.default}
\alias{Wasserstein.diss.distributional}
\alias{Wasserstein.diss.varietal}
\title{Wasserstein distance}
\usage{
Wasserstein.diss(x, ...)

\method{Wasserstein.diss}{default}(x, y, ...)

\method{Wasserstein.diss}{distributional}(x, log = FALSE, ...)

\method{Wasserstein.diss}{varietal}(x, package = "approxOT", ...)
}
\arguments{
\item{x}{the first sample as a vector}

\item{...}{optional arguments to the
\code{transport::wasserstein()} or
\code{approxOT::wasserstein()} functions.}

\item{y}{the second sample as a vector}

\item{log}{logical. Take the lograthm of the data before
calculating the distances?}

\item{package}{the name of the package that provides the 2D
Wasserstein distance. Currently, this can be either
\code{'transport'} or \code{approxOT}.}
}
\value{
a scalar value
}
\description{
Returns the Wasserstein distance between two samples
}
\examples{
data(Namib)
print(Wasserstein.diss(Namib$DZ$x[['N1']],Namib$DZ$x[['T8']]))
}
